analysis/pyproject.toml
[project]
name = "analysis"
version = "0.1.0"
description = "Analysis for the thesis 'Short-Term CO₂ Emissions Forecasting in Volatile Wind Power Grids using LSTM Neural Network'"
readme = "README.md"
requires-python = "~=3.12.9"
dependencies = [
    "energyquantified==0.14.4",
    "ipykernel==6.29.5",
    "matplotlib==3.10.1",
    "polars==1.26.0",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "requests==2.32.3",
    "scikit-learn==1.6.1",
    "statsmodels==0.14.4",
    "tensorflow==2.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["nbqa>=1.9.1", "ruff>=0.11.3"]

[tool.ruff]
line-length = 140

[tool.ruff.lint]
extend-select = [
    "I",   # isort
    "N",   # pep8-naming
    "ANN", # flake8-annotations
    "BLE", # flake8-blind-except
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
]
ignore = [
    "ANN002", # missing-type-args
    "ANN003", # missing-type-kwargs
    "N812",   # lowercase-imported-as-non-lowercase
    "N806",   # non-lowercase-variable-in-function
]

fixable = ["ALL"]
unfixable = [
    "F401", # unused import
]
